Preliminary support for 2.0 images.  These introduce a few new data types: 0xc0, 0xc1, 0xc3.  Thankfully they mirror to the existing values, 0x40, 0x41, 0x43.
For sanity checking purposes, ensure we're always dealing
with 0, 1 or 3.  If not, error out.  And that if we're
dealing with c, that the ROM major is 2.0.

The 2.0 images seem to introduce two unusual values in
frame class slots. 0x32 and 0x132.  These seem to map to
CodeBlock and CFunction.  So we'll do some swapping if
we aren't able to get back a class for the frame, and
we're in a 2.0 image.

To be useful for 2.0 images, we'll need to start reading in
the "high image" as well.